[ksqlDb.RestApi.Client]: split model builder-related types
