+++ ksqlDb.RestApi.Client/Metadata/FieldTypeBuilder.cs
@@ -1,0 +1,17 @@
+namespace ksqlDb.RestApi.Client.Metadata
+{
+  public interface IFieldTypeBuilder<TProperty>
+  {
+    internal IFieldTypeBuilder<TProperty> Ignore();
+  }
+
+  internal class FieldTypeBuilder<TProperty>(FieldMetadata fieldMetadata)
+    : IFieldTypeBuilder<TProperty>
+  {
+    public IFieldTypeBuilder<TProperty> Ignore()
+    {
+      fieldMetadata.Ignore = true;
+      return this;
+    }
+  }
+}
